ref: Move error handling related files into own module
