+++ ankihub/errors/__init__.py
@@ -1,0 +1,5 @@
+from .error_reporting import (  # noqa: F401
+    report_exception_and_upload_logs,
+    upload_logs_in_background,
+)
+from .errors import setup_error_handler  # noqa: F401
